feature(TP-181): add login xml, change nav graph, change mainActivity xml
