+++ app/src/main/java/com/example/reflect/presentation/adapters/ViewPager2Adapter.kt
@@ -3,14 +3,14 @@ package com.example.reflect.presentation.adapters
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.example.reflect.presentation.friends.fragment.FriendsFragment
-import com.example.reflect.presentation.profile.fragment.ProfileFragment
-import com.example.reflect.presentation.records.fragment.RecordsFragment
-import com.example.reflect.presentation.statistics.fragment.StatisticsFragment
+import com.example.reflect.presentation.screens.friends.fragment.FriendsFragment
+import com.example.reflect.presentation.screens.profile.fragment.ProfileFragment
+import com.example.reflect.presentation.screens.records.fragment.RecordsFragment
+import com.example.reflect.presentation.screens.statistics.fragment.StatisticsFragment
 
-class ViewPager2Adapter(
-    fragmentActivity: FragmentActivity
-) : FragmentStateAdapter(fragmentActivity) {
+class MainFragmentViewPagerAdapter(
+    fragment: Fragment
+) : FragmentStateAdapter(fragment) {
     override fun getItemCount(): Int = 4
 
     override fun createFragment(position: Int): Fragment =
