+++ app/src/main/java/com/example/reflect/presentation/screens/friends/fragment/FriendsFragment.kt
@@ -1,22 +1,22 @@
-package com.example.reflect.presentation.profile.fragment
+package com.example.reflect.presentation.screens.friends.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.reflect.databinding.FragmentProfileBinding
+import com.example.reflect.databinding.FragmentFriendsBinding
 
-class ProfileFragment : Fragment() {
+class FriendsFragment : Fragment() {
 
-    private var _binding: FragmentProfileBinding? = null
+    private var _binding: FragmentFriendsBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
-        _binding = FragmentProfileBinding.inflate(inflater, container, false)
+        _binding = FragmentFriendsBinding.inflate(inflater, container, false)
 
         return binding.root
     }
