+++ app/src/main/java/com/example/reflect/presentation/screens/login/fragment/LoginFragment.kt
@@ -1,0 +1,35 @@
+package com.example.reflect.presentation.screens.login.fragment
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.navigation.fragment.findNavController
+import com.example.reflect.R
+import com.example.reflect.databinding.FragmentLoginBinding
+
+class LoginFragment : Fragment() {
+
+    private var _binding: FragmentLoginBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View? {
+        _binding = FragmentLoginBinding.inflate(inflater, container, false)
+
+        binding.loginButton.setOnClickListener {
+            findNavController().navigate(R.id.mainFragment)
+        }
+
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+}
