+++ app/src/main/res/layout/activity_main.xml
@@ -1,49 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout
+<FrameLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".presentation.MainActivity">
+    android:layout_height="match_parent">
 
-    <com.google.android.material.bottomappbar.BottomAppBar
-        android:id="@+id/bottomAppBar"
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/mainFragmentContainer"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom"
-        android:padding="0dp"
-        app:backgroundTint="@color/surfaceContainer">
+        android:layout_height="match_parent"
+        android:name="androidx.navigation.fragment.NavHostFragment"
+        app:defaultNavHost="true"
+        app:navGraph="@navigation/navigation_graph" />
 
-        <com.google.android.material.bottomnavigation.BottomNavigationView
-            android:id="@+id/bottomNavBar"
-            android:layout_width="match_parent"
-            android:layout_height="75dp"
-            android:layout_marginEnd="16dp"
-            android:background="@android:color/transparent"
-            app:labelVisibilityMode="labeled"
-            app:elevation="0dp"
-            app:menu="@menu/bottom_nav_menu" />
-
-    </com.google.android.material.bottomappbar.BottomAppBar>
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/fabAI"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:contentDescription="@string/app_name"
-        android:layout_marginEnd="16dp"
-        android:layout_marginBottom="91dp"
-        app:shapeAppearanceOverlay="@style/CircleFab"
-        app:layout_anchor="@id/bottomAppBar"
-        app:layout_anchorGravity="top|end"
-        app:srcCompat="@drawable/ic_fab_ai" />
-
-
-    <androidx.viewpager2.widget.ViewPager2
-        android:id="@+id/mainActivityViewPager"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
+</FrameLayout>
