+++ app/src/main/res/layout/fragment_friends.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.friends.fragment.FriendsFragment">
+    tools:context=".presentation.screens.friends.fragment.FriendsFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/fragmentFriendsToolbar"
