+++ app/src/main/res/layout/fragment_profile.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.profile.fragment.ProfileFragment">
+    tools:context=".presentation.screens.profile.fragment.ProfileFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/fragmentProfileToolbar"
