+++ app/src/main/res/layout/fragment_records.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.records.fragment.RecordsFragment">
+    tools:context=".presentation.screens.records.fragment.RecordsFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/fragmentRecordsToolbar"
