+++ app/src/main/res/layout/fragment_statistics.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.statistics.fragment.StatisticsFragment">
+    tools:context=".presentation.screens.statistics.fragment.StatisticsFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/fragmentStatisticToolbar"
