+++ app/src/main/res/values/strings.xml
@@ -15,4 +15,14 @@
     <string name="contentDescriptionToolbarStreakIcon">Иконка стрика</string>
     <string name="contentDescriptionToolbarSearchIcon">Иконка поиска</string>
     <string name="contentDescriptionToolbarSettingsIcon">Иконка настроек</string>
+
+    <!-- Login Fragment    -->
+    <string name="loginTitle">Вход</string>
+    <string name="loginEmailHint">Почта</string>
+    <string name="loginPasswordHint">Пароль</string>
+    <string name="loginButton">Войти</string>
+    <string name="loginOr">или</string>
+    <string name="logintLikeGuest">Войти как гость</string>
+    <string name="logintRegistration">Регистрация</string>
+    <string name="loginForgotPassword">Забыли пароль?</string>
 </resources>
